+++ reductstore/src/storage/query/condition/operators.rs
@@ -1,4 +1,5 @@
 // Copyright 2024 ReductSoftware UG
 // Licensed under the Business Source License 1.1
 
+pub(crate) mod comparison;
 pub(crate) mod logical;