+++ reductstore/src/storage/query/condition/operators/logical/none_of.rs
@@ -2,7 +2,7 @@
 // Licensed under the Business Source License 1.1
 
 use crate::storage::query::condition::value::Value;
-use crate::storage::query::condition::{BoxedNode, Context, Node};
+use crate::storage::query::condition::{Boxed, BoxedNode, Context, Node};
 use reduct_base::error::ReductError;
 
 /// A node representing a logical NOT or NONE_OF operation.
@@ -27,13 +27,15 @@ impl Node for NoneOf {
     }
 }
 
+impl Boxed for NoneOf {
+    fn boxed(operands: Vec<BoxedNode>) -> Result<BoxedNode, ReductError> {
+        Ok(Box::new(Self::new(operands)))
+    }
+}
+
 impl NoneOf {
     pub fn new(operands: Vec<BoxedNode>) -> Self {
         Self { operands }
-    }
-
-    pub fn boxed(operands: Vec<BoxedNode>) -> BoxedNode {
-        Box::new(Self::new(operands))
     }
 }