+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - RS-528: Conditional query engine, [PR-640](https://github.com/reductstore/reductstore/pull/640)
 - RS-524: Check naming convention before creating or renaming entry, [PR-650](https://github.com/reductstore/reductstore/pull/650)
 - RS-545: Implement logical operators in conditional query, [PR-656](https://github.com/reductstore/reductstore/pull/656)
+- RS-529: Implement comparison operators in conditional query, [PR-662](https://github.com/reductstore/reductstore/pull/662)
 
 ### Changed